Added new project for Version 2.0 of the quiz
A new project has been created in Visual Studio. It takes the code from the first version of the program and starts to segment it into classes. This includes creating a model for the player, and services for selecting the difficulty, outputting the high scores and outputting the score of the quiz just done.

A dashboard will need to be created. This will include options such as taking a quiz, outputting the user's high scores, or exiting the program. The dashboard will be shown after the user has entered their name and the ASCII art is outputted.